+++ docker-compose.yaml
@@ -1,6 +1,18 @@
-version: '3'
+version: '3.8'
 
 services:
+
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "8080:8080"
+    environment:
+      - GO_ENV=production
+    depends_on:
+      - postgres
+
   postgres:
     image: postgres:latest
     container_name: postgres_db
