+++ go.mod
@@ -6,5 +6,5 @@ require github.com/lib/pq v1.10.9
 
 require (
 	github.com/golang-jwt/jwt/v5 v5.2.1
-	github.com/google/uuid v1.6.0 
+	github.com/google/uuid v1.6.0
 )
