+++ internals/server/repositories/token_repository.go
@@ -3,7 +3,7 @@ package repositories
 import (
 	"context"
 	"fast_food_auth/db"
-	"fast_food_auth/internals/models"
+	"fast_food_auth/internals/server/models"
 )
 
 type TokenRepository interface {
