+++ pkg/validation/validation_service.go
@@ -1,4 +1,4 @@
-package services
+package validation
 
 import (
 	"reflect"
